Allow {% if VALUE1 == VALUE2 %}, with two defines
I was surprised to see that blogc doesn't support testing for equality
between two defined values, it just supports comparison between a
variable and a constant string.  I want to be able to compare two
variables so I can build the same source with different "-D" argument on
the command line to produce different outputs.

This patch adds support for this pattern.  I changed the parser to
include the '"' at the beginning and end of a string, which allows the
renderer to determine if the user passed in a string or a variable name.
This is a bit hacky and causes some of the tests to fail -- these tests
look at the string values coming out of the parser.  I updated the tests
to match the new behavior.