content-parser: fixed DESCRIPTION variable.
it is now built as a single line variable, that contains the full
unparsed content of the first paragraph found in the source file,
instead of just the first line of it.

this also fixes a bug that prevented creating of DESCRIPTION variable,
if the source file contained only a single line paragraph.