blogc: add cli option to read list of source files from stdin
this patch adds `-i` option to command line. it will instruct blogc to
read stdin and parse it as a file where each source file is a line and
empty lines and lines starting with `#` are ignored.

this patch makes it possible to use blogc to build big blogs with lots
of pages, that would hit the operating system max command line length
when calling blogc to build pages.