content-parser: source-parser: extract TITLE from content header
this patch implements support to using the first header found in source
file as the TITLE variable. please note that if the TITLE variable is
defined on the source file's variable section it takes precedence.

this patch changes the old behaviour and can break some users' websites.
if you have some '{% ifdef TITLE %}' blocks in your template, they will
evaluate to true if you don't defined TITLE manually, but have a header
in your content.