make: use permalinks as atom ids
I have just realized that was implementing atom feeds wrong for long
time. atom ids are supposed to be valid uris. this is a small but
breaking change. it is supposed to make some clients think that all
of the old posts were republished. This can be avoided by running
something like this, before upgrading:

    $ blogc-make atom_dump > templates/atom.tmpl

however, keep in mind that your atom ids are invalid :/