blogc: generate table of contents tree
this commit allows users to use the `{{ TOCTREE }}` variable in their
templates, to get automatically generated table of contents for entries.
The variable is binded to each entry, so it can be used in any block
that runs in entry context, like `{% block entry %}`.

There's a small performance penalty for this, because the table of
contents is rendered for any entry, despite being used or not, and not
generated on-demand.

still missing documentation. tests are good enough.